# Major performance improvements on browsing rows