I made some major improvements to the performance that browsing rows had.

Using: xhprof

Ref: #16005

Explained:
- The first change is to just move a variable to avoid it to be re-built
- The second one is splitting in 3 functions the description array to avoid wasting time fetching translations that will not be used (99% are not used, only one is picked up) (yes I built a script to check all the keys and values exist and are the same than before, diff should be okay)

Dataset and method:
- A table of more that 1.1 Gio of data
- 7 973 078 rows
- show rows by 500 on browse page

### Before (14 229ms)
![image](https://user-images.githubusercontent.com/7784660/105772629-c7e46580-5f62-11eb-943b-cdef51c15157.png)

#### Call refresh button (14 981ms)
![image](https://user-images.githubusercontent.com/7784660/105772701-e0ed1680-5f62-11eb-9742-33c087d3b3f1.png)

### After (7 524 ms)
![image](https://user-images.githubusercontent.com/7784660/105772797-0a0da700-5f63-11eb-9ad0-a1db5e7f50a2.png)

#### Call refresh button (6 673 ms)
![image](https://user-images.githubusercontent.com/7784660/105772878-2873a280-5f63-11eb-9c02-30bdf47b9d3f.png)