@ibennetch 2021-01-28
I'm very impressed with the speed improvements you've made here. Nice work!