@williamdes 2021-01-28
> I'm very impressed with the speed improvements you've made here. Nice work!

Thank you @ibennetch !

> Really nice. 1st_place_medal:

Thank you @mauriciofauth !

> I'd like to get this in 5.1 before I release 5.1.0rc2.

I will merge that right now, could you please review #16584 ?


NB: some discussion is going on at https://github.com/code-lts/welcome/issues/6 about xhprof, I intend to make it usable for other people so our users can send us xhprof files